:sparkles: Add UserDefaults+WalletDataStoreProtocol.swift
